% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdb.R
\docType{data}
\name{.SAMADB_ID}
\alias{.SAMADB_ID}
\alias{.SAMADB_T}
\title{Global Identifier Macros}
\usage{
.SAMADB_ID
.SAMADB_T
}
\description{
The macro \code{.SAMADB_ID} contains the string \code{c("dsid", "series")} denoting
the names of ID variables that identify the cross-sectional dimension in the database. All series codes are unique across datasets.

The macro \code{.SAMADB_T} contains the string
\code{c("date", "year", "quarter", "month", "day")} denoting temporal identifiers generated by \code{\link{sm_expand_date}}.
The \code{"date"} variable is sufficient to uniquely identify a point in time in the database.

Each value in the database is uniquely identified by dsid, series and date.
}
\examples{
.SAMADB_ID
.SAMADB_T


}
\seealso{
\code{\link{samadb}}
}
\keyword{datasets}
