% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samadb.R
\docType{package}
\name{samadb-package}
\alias{samadb-package}
\alias{samadb}
\title{South Africa Macroeconomic Database API}
\description{
An R API providing access to a relational database with public macroeconomic data for South Africa, obtained from
from the South African Reserve Bank (SARB) and Statistics South Africa (STATSSA), and updated on a regular basis via the
EconData (https://www.econdata.co.za/) platform and automated scraping of the SARB and STATSSA websites.
The database is maintained at the Department of Economics at Stellenbosch University.
}
\section{Functions}{

Functions and data providing information about the available data

\code{\link[=sm_datasources]{sm_datasources()}}\cr
\code{\link[=sm_datasets]{sm_datasets()}}\cr
\code{\link[=sm_series]{sm_series()}}

Function to retrieve the data from the database

\code{\link[=sm_data]{sm_data()}}

Functions to reshape data and add temporal identifiers

\code{\link[=sm_pivot_wider]{sm_pivot_wider()}}\cr
\code{\link[=sm_pivot_longer]{sm_pivot_longer()}}\cr
\code{\link[=sm_expand_date]{sm_expand_date()}}

Function to export wide format data to Excel

\code{\link[=sm_write_excel]{sm_write_excel()}}\cr

Helper functions to convert inputs to R dates and transpose the data

\code{\link[=sm_as_date]{sm_as_date()}}\cr
\code{\link[=sm_transpose]{sm_transpose()}}\cr

Global Macros with core ID variables in the database

\code{\link{.SAMADB_ID}}\cr
\code{\link{.SAMADB_T}}
}

